refactor desc+attachment fields into component + tests
