+++ src/components/EntryEditor/Description/index.js
@@ -1,0 +1,4 @@
+import Description from "./Description";
+
+export { Description };
+export default Description;
